[LEARN] next: Lesson 1 from Next.js series
This commit demonstrates the difference between server and client
components in Next.js.

The `Greet` component is a server component that is rendered server-side
and displayed in the browser. However, you cannot interact with it.
To interact with the component, you need to use a client component,
which in this case is the `Counter` component.
In the `Counter` component, you can see that the state is maintained and
the value is updated on the client side using the `useState` hook.

In the `useState` hook, the value passed as an argument is the initial
value. For example:

`const [count, setCount] = useState(0);` will start the counter from 0.

While

`const [count, setCount] = useState(1);` will start the counter from 1.