[LEARN] nextjs: data fetching
In this commit, we have learned how to use server-side data fetching as
well as client-side data fetching in Next.js.

For client-side data fetching, we handle the loading and error states
manually. This is demonstrated in the `users-client` folder. We have
explicitly added a 2-second delay to demonstrate the loading component.

For server-side data fetching, we don't have to handle the loading and
error states explicitly. Instead, we create components for loading and
error states. For example, in the `users-server` folder, we have created
`loading.tsx` and `error.tsx`.

URLs for testing are as follows:
- http://localhost:3000/users-client/
- http://localhost:3000/users-server/